yad2k-em3d/requirements.txt
absl-py==0.1.11
alabaster==0.7.10
ansicolors==1.1.8
ansiwrap==0.8.3
appdirs==1.4.3
astor==0.6.2
astroid==1.6.3
Babel==2.5.3
backcall==0.1.0
bleach==1.5.0
certifi==2018.1.18
chainmap==1.0.2
chardet==3.0.4
cloudpickle==0.5.2
combomethod==1.0.10
cycler==0.10.0
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imagesize==1.0.0
ipdb==0.11
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.0
mccabe==0.6.1
mementos==1.2.8
mistune==0.8.3
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nibabel==2.2.1
nulltype==2.2.11
numeral==0.1.0.11
numpy==1.14.1
numpydoc==0.8.0
opencv-python==3.4.0.12
options==1.4.8
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2
psutil==5.4.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
pymrt==0.0.2.1
PyOpenGL==3.1.0
pyparsing==2.2.0
PyQt5==5.9.2
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
readline==6.2.4.1
requests==2.18.4
rope==0.10.7
say==1.6.5
scikit-image==0.13.1
scipy==1.0.0
seaborn==0.8.1
show==1.6.0
simplegeneric==0.8.1
simplere==1.2.13
sip==4.19.8
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
sympy==1.1.1
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
testpath==0.3.1
textdata==1.7.3
textwrap3==0.9.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
